Extract separate modules

+++ src/player.rs
@@ -1,8 +1,8 @@
 use crate::constants::*;
-use avian2d::math::{AdjustPrecision, Scalar, Vector};
+use crate::shared::{Player, WrapAroundEntity};
+use avian2d::math::Vector;
 use avian2d::prelude::*;
 use bevy::asset::RenderAssetUsages;
-use bevy::color::palettes::tailwind;
 use bevy::mesh::{Indices, PrimitiveTopology};
 use bevy::prelude::*;
 
@@ -11,18 +11,8 @@ pub struct PlayerPlugin;
 impl Plugin for PlayerPlugin {
   fn build(&self, app: &mut App) {
     app
-      .add_message::<InputAction>()
       .add_systems(Startup, spawn_player_system)
-      .add_systems(
-        Update,
-        (
-          keyboard_input_system,
-          movement_system,
-          wraparound_system,
-          render_gizmos_system,
-          (update_snake_body_system, refresh_snake_mesh_system).chain(),
-        ),
-      );
+      .add_systems(Update, ((update_snake_body_system, refresh_snake_mesh_system).chain(),));
   }
 }
 
@@ -49,26 +39,17 @@ impl Controller {
   }
 }
 
-/// A marker component for the player entity.
 #[derive(Component)]
-struct Player;
-
-/// A marker component for entities that should wrap around the screen edges.
-#[derive(Component)]
-struct WrapAroundEntity;
-
-/// A [`Message`] written for an input action.
-#[derive(Message)]
-enum InputAction {
-  Move(Scalar),
-  Action,
-}
-
-#[derive(Component)]
-struct SnakeBody {
+pub(crate) struct SnakeBody {
   positions: Vec<Vec2>,
   distance_since_last_sample: f32,
   segment_entities: Vec<Entity>,
+}
+
+impl SnakeBody {
+  fn len(&self) -> usize {
+    self.positions.len()
+  }
 }
 
 #[derive(Component)]
@@ -110,7 +91,7 @@ fn spawn_player_system(
     WrapAroundEntity,
     Sprite::from_image(snake_head_handle),
     SnakeBody::default(),
-    Controller::new(Collider::circle(5.)),
+    Controller::new(Collider::circle(SNAKE_HEAD_SIZE)),
     Transform::default(),
     Friction::ZERO.with_combine_rule(CoefficientCombine::Min),
     Restitution::ZERO.with_combine_rule(CoefficientCombine::Min),
@@ -130,81 +111,6 @@ fn spawn_player_system(
   ));
 }
 
-fn render_gizmos_system(mut gizmos: Gizmos, player_query: Query<&Transform, With<SnakeBody>>) {
-  for transform in player_query.iter() {
-    gizmos.circle_2d(
-      Isometry2d::from_translation(Vector::new(transform.translation.x, transform.translation.y)),
-      1.,
-      Color::from(tailwind::AMBER_400),
-    );
-  }
-}
-
-/// Sends [`InputAction`] events based on keyboard input.
-fn keyboard_input_system(
-  mut input_action_writer: MessageWriter<InputAction>,
-  keyboard_input: Res<ButtonInput<KeyCode>>,
-) {
-  let left = keyboard_input.any_pressed([KeyCode::KeyA, KeyCode::ArrowLeft]);
-  let right = keyboard_input.any_pressed([KeyCode::KeyD, KeyCode::ArrowRight]);
-  let horizontal = right as i8 - left as i8;
-  let direction = horizontal as Scalar;
-  if direction != 0.0 {
-    input_action_writer.write(InputAction::Move(direction));
-  }
-  if keyboard_input.just_pressed(KeyCode::Space) {
-    input_action_writer.write(InputAction::Action);
-  }
-}
-
-/// Responds to [`InputAction`] events and moves character controllers accordingly.
-fn movement_system(
-  time: Res<Time>,
-  mut input_action_messages: MessageReader<InputAction>,
-  mut controllers: Query<(&Transform, &mut LinearVelocity, &mut AngularVelocity), With<Player>>,
-) {
-  let delta_time = time.delta_secs_f64().adjust_precision();
-  for (transform, mut linear_velocity, mut angular_velocity) in &mut controllers {
-    let mut has_movement_input = false;
-    let direction = (transform.rotation * Vec3::Y).normalize_or_zero();
-    let velocity = direction * MOVEMENT_SPEED;
-    linear_velocity.x = velocity.x;
-    linear_velocity.y = velocity.y;
-
-    for event in input_action_messages.read() {
-      has_movement_input = true;
-      match event {
-        InputAction::Move(direction) => {
-          angular_velocity.0 = -*direction * ROTATION_SPEED * delta_time;
-        }
-        InputAction::Action => {
-          debug!("[Not implemented] Action received");
-        }
-      }
-    }
-    if !has_movement_input {
-      angular_velocity.0 = 0.;
-    }
-  }
-}
-
-/// Wraps the relevant entities around the screen edges, making them reappear on the opposite side.
-fn wraparound_system(mut entities: Query<&mut Transform, With<WrapAroundEntity>>) {
-  let extents = Vec3::new(RESOLUTION_WIDTH as f32 / 2., RESOLUTION_HEIGHT as f32 / 2., 0.);
-  for mut transform in entities.iter_mut() {
-    if transform.translation.x > (extents.x + WRAPAROUND_MARGIN) {
-      transform.translation.x = -extents.x - WRAPAROUND_MARGIN;
-    } else if transform.translation.x < (-extents.x - WRAPAROUND_MARGIN) {
-      transform.translation.x = extents.x + WRAPAROUND_MARGIN;
-    }
-    if transform.translation.y > (extents.y + WRAPAROUND_MARGIN) {
-      transform.translation.y = -extents.y - WRAPAROUND_MARGIN;
-    } else if transform.translation.y < (-extents.y - WRAPAROUND_MARGIN) {
-      transform.translation.y = extents.y + WRAPAROUND_MARGIN;
-    }
-  }
-}
-
 fn update_snake_body_system(
   mut commands: Commands,
   mut player_query: Query<(&Transform, &mut SnakeBody), With<Player>>,
@@ -212,6 +118,14 @@ fn update_snake_body_system(
   mut materials: ResMut<Assets<ColorMaterial>>,
 ) {
   for (transform, mut body) in &mut player_query {
+    if body.len() > MAX_CONTINUOUS_SNAKE_LENGTH + SNAKE_GAP_LENGTH {
+      // TODO: Create new body
+    }
+
+    if body.len() > MAX_CONTINUOUS_SNAKE_LENGTH {
+      continue;
+    }
+
     let current_position = transform.translation.truncate() - (transform.rotation * Vec3::Y * 10.).truncate();
 
     if body.positions.is_empty() {
@@ -251,17 +165,17 @@ fn refresh_snake_mesh_system(
   mut mesh_query: Query<&mut Mesh2d, With<SnakeTail>>,
   mut meshes: ResMut<Assets<Mesh>>,
 ) {
-  // for body in &player_query {
-  //   if body.positions.len() < 2 {
-  //     continue;
-  //   }
-  //
-  //   for mesh_handle in &mut mesh_query {
-  //     if let Some(mesh) = meshes.get_mut(&mesh_handle.0) {
-  //       *mesh = create_snake_body_mesh(&body.positions);
-  //     }
-  //   }
-  // }
+  for body in &player_query {
+    if body.positions.len() < 2 {
+      continue;
+    }
+
+    for mesh_handle in &mut mesh_query {
+      if let Some(mesh) = meshes.get_mut(&mesh_handle.0) {
+        *mesh = create_snake_body_mesh(&body.positions);
+      }
+    }
+  }
 }
 
 fn create_snake_body_mesh(positions: &[Vec2]) -> Mesh {
